Create angry_check.py
There are two friends, John and Smith, and the parameters j_angry and s_angry to indicate if each is angry. You are in trouble if both of them are angry or no one of them is angry.

Now, complete the function which returns true if you are in trouble, else return false

Example 1:

Input: j_angry = True, s_angry = True Output: True Explanation: Since both of them are angry, so you are in trouble.